Modularize quell-client
