+++ quell-client/createQueryStr.js
@@ -1,0 +1,29 @@
+function createQueryStr(queryObject) {
+  const openCurl = ' { ';
+  const closedCurl = ' } ';
+
+  let mainStr = '';
+
+  for (let key in queryObject) {
+    mainStr += key + openCurl + stringify(queryObject[key]) + closedCurl;
+  }
+
+  function stringify(fieldsArray) {
+    let innerStr = '';
+    for (let i = 0; i < fieldsArray.length; i++) {
+      if (typeof fieldsArray[i] === 'string') {
+        innerStr += fieldsArray[i] + ' ';
+      }
+      if (typeof fieldsArray[i] === 'object') {
+        for (let key in fieldsArray[i]) {
+          innerStr += key + openCurl + stringify(fieldsArray[i][key]);
+          innerStr += closedCurl;
+        }
+      }
+    }
+    return innerStr;
+  }
+  return openCurl + mainStr + closedCurl;
+};
+
+export default createQueryStr
